Add x, y, and z shorthands to id and range classes
The numbering of dimensions in SYCL is aligned with the numbering of dimensions
in ISO C++, such that the highest-numbered dimension is the fastest-moving.

This numbering is inconvenient when working with generic functions compatible
with one-, two-, or three-dimensional ranges, since developers must account
somehow for differences in the numbering of the dimensions. One common solution
is to define helper functions called x(), y() and z() to encapsulate this
logic. This solution can also assist with the migration of code from other
languages (e.g., OpenCL and CUDA), and may provide a simpler mental model for
SYCL developers working with images or other forms of graphics interop.

This commit adds x(), y() and z() functions directly to the id and range
classes, providing a consistent way for SYCL developers to use this indexing
pattern, rather than relying on each SYCL code base to define and maintain
compatible index abstractions.