Revert "Merge pull request #698 from gmlueck/gmlueck/cherry-pick-693"
Revert moving the Acknowledgements chapter.  Although we want to move
this chapter, we don't want to move it as a revision to SYCL 2020
because it causes all the other chapter numbers to change.  Therefore
keep this change in the "main" branch, but remove it from "sycl-2020".

This reverts commit 665315e968b8e79562a752338c976bbd248cd313, reversing
changes made to 1399526a058c12cdc72d84edbbecf503e299c8c6.