Tweak wording about ADL namespaces
The previous wording said that the namespace containing the
`__writeable_swizzle__` \ `__const_swizzle__` type must be on the ADL
search set.  However, these types aren't defined by the spec, so their
namespace isn't defined.  Change the wording to say that it is the
`sycl` namespace that must be on the ADL search set.