Tweak ADL wording again
On second thought, we don't want the `sycl` namespace on the ADL search
set.  The swizzle operators are defined as friend function *inside* the
class, so it doesn't help to put the `sycl` namespace on the search set.
It's better if the implementation defines the swizzle type such that
`sycl` is not on the search set (e.g. by defining it in a nested
namespace).