Rename khr::work_item to khr::member_item
Several people had suggested that allowing "work-item" to mean both something
similar to "thread of execution" (e.g., as in "this work-item", "the calling
work-item") and an item within a group had the potential to be confusing.

"member_item" more clearly describes an item that is a member of a group,
and cannot be confused with the more general concept of a work-item.