Clarify the meaning of different device types
The description of what constitutes a CPU or GPU has always been unclear.

This commit adjusts these descriptions to clarify that it is ultimately up to
the implementation to decide whether a device is identified as a specific
device type. It also attempts to separate normative behaviors (like the
connection between device type aspects and info::device_type) from
non-normative descriptions of what a device type may mean.