Add issue templates
Managing and automating our handling of GitHub issues would be easier if we
imposed more consistent formatting and encouraged contributors to provide
specific information.

This commit introduces three different templates:

- Bug Report: Report a bug in the specification.
- Clarification Request: Highlight that something is unclear.
- Feature Request: Request a new SYCL feature/extension.

We may require additional issue types in the future, but based on recent
efforts to reduce the GitHub backlog, it seems like the majority of issues
fall into one of these categories.