Adjust constexpr behavior of group interface
The group interface (which we hope to replace with a concept) should not
dictate whether or not a function is constexpr unless all groups must have the
same behavior.

Since we have two existing groups (work-group and sub-group) where it is
impossible to determine their sizes at compile-time, it would be misleading to
declare functions relating to their size as constexpr.

Making this change to the interface requires the member functions to be
duplicated for every class. I've used this as an opportunity to clarify the
behavior of these functions in a way that is specific to each class (e.g.,
highlighting that a sub-group's indices are with respect to its parent
work-group).