Clarify free is not guaranteed to be non-blocking
The previous wording about "not waiting" could have been read as a guarantee
that calls to sycl::free are always non-blocking. In practice, calls to
sycl::free may need to block until the device can satisfy the deallocation
request (depending on implementation details tied to specific backends).

Although the new wording cannot be read this way, adding a non-normative note
may help developers to understand that they should not rely on the behavior of
specific implementations.