Remove "inter-device transfer" from copy rules
The previous wording said that the effect of various copy and move constructors
must be equivalent to a bitwise copy only when performing an inter-device
transfer. However, there is no way for a developer to detect when an
inter-device transfer is happening.

The real requirement here is stronger: the effect of each of these copy and
move constructors must *always* be equivalent to a bitwise copy.