Clarify specializing sycl:: templates is invalid
ISO C++ allows developers to be able to define specializations of class
templates in the std:: namespace unless prohibited, but this has caused issues
in the past.

Rather than follow the ISO C++ precedent, the SYCL WG has decided to
consider things on a case-by-case basis and explicitly permit specialization
of class templates in the sycl:: namespace only where there is a strong
use-case for doing so.

Initially, it is only valid to specialize the following class templates:
- known_identity
- has_known_identity
- is_device_copyable