Clarify thread-safety guarantees
Previously, the thread-safety guarantee only applied to member functions of
SYCL classes and did not apply to free functions, function templates, etc.
This change clarifies that the thread-safety guarantee applies to all
functions.

A new non-normative note further clarifies that this guarantee is not intended
to eliminate data races from SYCL programs entirely.