Deprecate the half_precision:: namespace
The functions in the half_precision namespace correspond to math functions with
the half_ prefix in OpenCL. In OpenCL (and SYCL 1.2.1), the functions have a
guaranteed minimum precision, but in SYCL 2020 the precision is
backend-specific and non-portable.

Deprecating these functions in favor of using real half precision (via
sycl::half or std::float16_t) will make SYCL simpler.