Explicitly delete handler move/copy constructors
Previously we said in words that a handler could not be moved or copied,
but it was unclear whether this meant that it should be forbidden by the class
implementation or simply resulted in undefined behavior.

This change explicitly deletes the handler move and copy constructors, making
it clear that it is the implementation's responsibility to prevent a handler
from being moved or copied.