Remove all handler overloads
A more detailed investigation of performance overheads in SYCL implementations
has uncovered that the cost associated with using a handler is similar to the
cost associated with returning a sycl::event.

This commit removes all the handler overloads from the KHR, as a first step
towards introducing an alternative design that does not depend on handler.