Allow kernel_bundle as a requirement
Adding a kernel_bundle<executable> as a requirement should have the same effect
as calling handler::use_kernel_bundle.

handler::use_kernel_bundle cannot be used in conjunction with commands
accepting a kernel object, and the command is defined as ignoring the
kernel_bundle in that case.

Since the new API receives the command and all requirements simultaneously, and
the requirements are known at compile-time, we can instead use a Constraint to
ensure that such code doesn't compile.