Add note about work-item scope atomics
In OpenCL, these atomics are only required to support a very specific use-case
involving images, and are forbidden in all other contexts.

In SYCL, we would like a work-item to be viewed as a degenerate case of a group
containing a single work-item. Work-item scope atomics should thus be
permitted, and their effect should be equivalent to non-atomic operations.