Clarify behavior of atomic_ref DefaultScope
An implementation of atomic_ref<T> that was not lock-free needs to know which
work-items may access the lock in order to decide where to allocate the lock.

This additionally serves as a clarification of the behavior of work-item scope;
using an atomic_ref with work-item scope at the same time as an atomic_ref
with broader scope is invalid.