Clarify that using different scopes is a data race
Even when using two atomic_ref objects with the same DefaultScope, it's
possible to encounter a data race by overriding the scope parameter of
individual operations.

This is a general clean-up but was motivated by work-item scope atomics:
any potentially concurrent use of work-item scope atomics and atomics with a
different scope results in undefined behavior.