Clarify which atomic operations synchronize
The ISO C++ synchronizes-with relationship does not account for scopes.
The scopes do not need to match exactly, but there are restrictions on which
pairs of scopes are valid.

This is the final part of the clarification for work-item scope atomics;
a work-item scope atomic cannot sychronize with the atomic operations
performed by other work-items, and so their effects are not guaranteed to be
visible to other work-items without some other synchronization taking place.