Add new requirements for sub-group sizes
Since the introduction of SYCL 2020 there have been multiple requests from
developers and implementers to provide more detail about how sub-groups are
expected to work.

This change is a first step towards more portable sub-groups, by ensuring that
sub-group behavior is well-defined in cases where the highest-numbered
dimension of a work-group is divisible by the sub-group size. Dividing the
highest-numbered dimension into sub-groups is the obvious thing to do in this
case, and matches the behavior of existing implementations.

The way in which other work-groups are divided into sub-groups remains
implementation-defined, but this change introduces a non-normative note to hint
at possible implementation choices. Eventually we may want to consider exposing
these choices via device queries or similar, and implementations are exploring
this via extensions.