Clarify effect of using accessors outside commands
The previous wording said that certain accessor member functions may only be
called from within a command, but did not say what should happen if this
condition was violated.

The new wording clarifies that this results in undefined behavior, giving
different implementations the freedom to decide how to diagnose this error.