Fix compare_exchange_* with only one memory_order
Previously, compare_exchange_weak and compare_exchange_strong with only one
memory_order were defined as being equivalent to passing that value as both
the success and failure arguments of the overload accepting two memory_orders.
This is a bug, because the overload accepting two memory_orders clearly states
that failure does not accept certain values of memory_order.

The definitions of the equivalent functions in ISO C++ do not have this issue,
and so this commit aligns the behavior of these functions with ISO C++.