Add note about vec behavior in relational funcs
Add a non-normative note which draws attention to the fact that the
relational functions taking `vec` inputs have "truthiness" semantics
that differ from from most other C++ functions.

The semantic of these functions is inherited from the SYCL 1.2.1
specification, and it aligns with the OpenCL C semantics of vectors.

Note: The change to "reflow.py" is necessary to make the script realize
that the period ending a sentence like "Blah blah OpenCL C." really is
the period at the end of a sentence.  Without this change, the script
thinks the "C." is a middle initial in someone's name, so it doesn't
interpret the period as an end-of-sentence marker.

Closes #902